pyproject.toml
[project]
name = "medscribe-ai"
version = "1.0.0"
description = "Agentic Clinical Documentation System powered by HAI-DEF models (MedGemma, MedASR, MedSigLIP)"
readme = "README.md"
requires-python = ">=3.12"
license = "CC-BY-4.0"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.45.0",
    "accelerate>=0.30.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.0.0",
    "Pillow>=10.0.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "gradio>=4.40.0",
    "bitsandbytes>=0.43.0",
    "requests>=2.31.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.5.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I"]
